Changed gain setting from int to float

+++ pyobs_asi/asicamera.py
@@ -201,7 +201,7 @@ class AsiCamera(BaseCamera, ICamera, IWindow, IBinning, IImageFormat, IAbortable
         self._camera.set_control_value(asi.ASI_EXPOSURE, int(exposure_time * 1e6))
 
         # set gain
-        self._camera.set_control_value(asi.ASI_GAIN, int(self._gain))
+        self._camera.set_control_value(asi.ASI_GAIN, float(self._gain))
 
         log.info(
             "Starting exposure with %s shutter for %s seconds and %s gain...", "open"
